feat: improved deck list info when decks are empty (#1564)
* REFAC: Refactored reminders into a manager

* REFAC:  Moved some functions to the review queue loader

* FEAT: Added a no decks info msg
